Auth: Update oauth2 mechanics.
Changes have happened in the devise and oauth libraries since the last
good deployment. Current builds will fail to enable signing in due to
a slight API change and the need to incroporate CSRF protection. As it
stands, users (developers) clicking the 'sign in with google link'
will see the arcane 'Not found. Authorization passthru.'

Add the method argument to ensure the button click triggers a post request.
Add the omniauth-rails_csrf_protection gem.

References #178, #140.